Sort mdl values list into (near-)alphabetical order
The list is now in alphabetical order, to make things easier to find, with the exception of `menu` which was push one row up, so that the two footer types would be adjacent to each other.

Why? Because I found I was scanning this list often, and its random order was a cognitive barrier to efficiency.